Move checkValidSocket to the unix test file
